+++ graphqlUploadExpress.mjs
@@ -9,13 +9,20 @@
  */
 
 import defaultProcessRequest from "./processRequest.mjs";
+import requestFinished from "./requestFinished.mjs";
 
 /**
  * Creates [Express](https://expressjs.com) middleware that processes incoming
  * [GraphQL multipart requests](https://github.com/jaydenseric/graphql-multipart-request-spec)
  * using {@linkcode processRequest}, ignoring non multipart requests. It sets
- * the request `body` to be similar to a conventional GraphQL POST request for
- * following GraphQL middleware to consume.
+ * the Express request property `body` to be similar to a conventional GraphQL
+ * POST request for following GraphQL middleware to consume. Also, it
+ * temporarily monkey patches the Express response method
+ * {@linkcode Response.send send} to first wait for the request to finish
+ * (either because it disconnects early, or it finishes uploading and the
+ * response can be sent), because sending a response before the request has
+ * finished uploading typically causes the HTTP client to error without
+ * processing the response.
  * @param {ProcessRequestOptions & {
  *   processRequest?: ProcessRequestFunction,
  * }} options Options.
@@ -44,27 +51,21 @@ export default function graphqlUploadExpress({
   ...processRequestOptions
 } = {}) {
   /**
-   * [Express](https://expressjs.com) middleware that processes incoming
-   * [GraphQL multipart requests](https://github.com/jaydenseric/graphql-multipart-request-spec)
-   * using {@linkcode processRequest}, ignoring non multipart requests. It sets
-   * the request `body` to be similar to a conventional GraphQL POST request for
-   * following GraphQL middleware to consume.
    * @param {Request} request Express request.
    * @param {Response} response Express response.
    * @param {NextFunction} next Invokes the next middleware.
    */
   function graphqlUploadExpressMiddleware(request, response, next) {
     if (request.is("multipart/form-data")) {
-      const requestEnd = new Promise((resolve) => request.on("end", resolve));
+      // Todo: Find a less hacky way that avoids monkey patching.
       const { send } = response;
-
-      // @ts-ignore Todo: Find a less hacky way to prevent sending a response
-      // before the request has ended.
       response.send = (...args) => {
-        requestEnd.then(() => {
+        requestFinished(request).then(() => {
           response.send = send;
           response.send(...args);
         });
+
+        return response;
       };
 
       processRequest(request, response, processRequestOptions)